requirements.txt
torch
torchvision
repackage
numpy
pandas
Pillow
tensorboardX
scipy
grad-cam